Implement new, simpler probability estimation method
This skips cross-validation and computes decision values using the original SVM,
and then uses exclusively the model-provided decision values, rather than the training
targets, to train the sigmoid. This is much faster than the original method (and can be
improved further as we currently train the main model twice), works better on smaller
and/or deduplicated datasets (no need to duplicate the training set), and is guaranteed to always predict the highest probability for the class the plain classifier
(without probability estimation) would predict.